+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * The plugin provides an ESM, CSJ and UMD version
 * The "class" option accepts a single class name or an array of class names
 * Sources were converted to TypeScript
+* replace deprecated properties "page{X,Y}Offset" with scroll{X,Y}
 
 ----